Moved ITunes namespace data to itunes package
