+++ src/main/kotlin/com/github/magneticflux/rss/Converters.kt
@@ -1,5 +1,7 @@
 package com.github.magneticflux.rss
 
+import com.github.magneticflux.rss.itunes.Explicit
+import com.github.magneticflux.rss.itunes.ITunesTopLevelCategory
 import org.simpleframework.xml.convert.Converter
 import org.simpleframework.xml.stream.InputNode
 import org.simpleframework.xml.stream.OutputNode
@@ -11,12 +13,12 @@ import java.util.Locale
 /**
  * A Persister that is used by various [Converter] objects.
  */
-private val fallbackPersister = createRssPersister()
+internal val fallbackPersister = createRssPersister()
 
 /**
  * Gets the prefix + name or just name if prefix is null.
  */
-private val InputNode.fullName: String
+internal val InputNode.fullName: String
     get() {
         val prefix = this.prefix
         val name = this.name
@@ -26,7 +28,7 @@ private val InputNode.fullName: String
 /**
  * Creates an iterator over an InputNode's children.
  */
-private val InputNode.children: Iterator<InputNode>
+internal val InputNode.children: Iterator<InputNode>
     get() {
         return InputNodeChildIterator(this)
     }
@@ -419,69 +421,5 @@ object SourceConverter : Converter<Source> {
     override fun write(node: OutputNode, value: Source) {
         node.setAttribute("url", URLTransform.write(value.url))
         node.value = value.text
-    }
-}
-
-/**
- * @author Mitchell Skaggs
- * @since 1.0.5
- */
-object ITunesTopLevelCategoryConverter : Converter<ITunesTopLevelCategory> {
-    override fun read(node: InputNode): ITunesTopLevelCategory {
-        val text = node.getAttribute("text").value
-        val subCategories = mutableListOf<ITunesSubCategory>()
-
-        node.children.forEach {
-            when (it.fullName) {
-                "itunes:category" -> subCategories += fallbackPersister.read(ITunesSubCategory::class.java, it)
-            }
-        }
-
-        return ITunesTopLevelCategory(text, subCategories)
-    }
-
-    override fun write(node: OutputNode, value: ITunesTopLevelCategory) {
-        node.setAttribute("text", value.text)
-        value.subCategories.forEach {
-            fallbackPersister.write(it, node)
-        }
-    }
-}
-
-/**
- * @author Mitchell Skaggs
- * @since 1.0.5
- */
-object ITunesSubCategoryConverter : Converter<ITunesSubCategory> {
-    override fun read(node: InputNode): ITunesSubCategory {
-        val text = node.getAttribute("text").value
-
-        return ITunesSubCategory(text)
-    }
-
-    override fun write(node: OutputNode, value: ITunesSubCategory) {
-        node.setAttribute("text", value.text)
-    }
-}
-
-/**
- * @author Mitchell Skaggs
- * @since 1.0.5
- */
-object ExplicitConverter : Converter<Explicit> {
-    override fun read(node: InputNode): Explicit {
-        val value: String? = node.value
-        return when {
-            value.equals("yes", true) -> Explicit.YES
-            value.equals("no", true) -> Explicit.NO
-            else -> throw IllegalStateException("Incorrect 'itunes:explicit' text $value is not 'yes' or 'no' ignoring case")
-        }
-    }
-
-    override fun write(node: OutputNode, value: Explicit) {
-        node.value = when (value.isExplicit) {
-            true -> "yes"
-            false -> "no"
-        }
     }
 }
