+++ src/main/kotlin/com/github/magneticflux/rss/Rss.kt
@@ -1,5 +1,7 @@
 package com.github.magneticflux.rss
 
+import com.github.magneticflux.rss.itunes.Explicit
+import com.github.magneticflux.rss.itunes.ITunesTopLevelCategory
 import org.simpleframework.xml.Attribute
 import org.simpleframework.xml.Element
 import org.simpleframework.xml.ElementList
@@ -292,48 +294,3 @@ data class Source @JvmOverloads constructor(
         @get:Text(required = false)
         val text: String? = null
 )
-
-/**
- * This class represents an itunes:category in a [Channel] or an [Item].
- *
- * @author Mitchell Skaggs
- * @since 1.0.5
- * @see ITunesTopLevelCategoryConverter
- */
-@Root(name = "category")
-@Namespace(reference = ITUNES_REFERENCE)
-data class ITunesTopLevelCategory(
-        val text: String,
-        val subCategories: List<ITunesSubCategory>
-)
-
-/**
- * This class represents an itunes:category in a [ITunesTopLevelCategory].
- *
- * @author Mitchell Skaggs
- * @since 1.0.5
- * @see ITunesSubCategoryConverter
- */
-@Root(name = "category")
-@Namespace(reference = ITUNES_REFERENCE)
-data class ITunesSubCategory(
-        val text: String
-)
-
-/**
- * This class represents an itunes:explicit in a [Channel] or an [Item].
- *
- * @author Mitchell Skaggs
- * @since 1.0.5
- * @see ExplicitConverter
- */
-@Root(name = "explicit")
-@Namespace(reference = ITUNES_REFERENCE)
-data class Explicit(
-        val isExplicit: Boolean
-) {
-    companion object {
-        val YES = Explicit(true)
-        val NO = Explicit(false)
-    }
-}
